+++ packages/valibot-form-data/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`valibot-form-data`](https://github.com/typed-web/typed-web/tree/main/packages/valibot-form-data). It follows [semantic versioning](https://semver.org/).
 
+## Unreleased
+
+- Simplify `formData` validator to use updated `preprocessFormData` from `@typed-web/form-utils`
+
 ## v0.4.0 (2026-02-05)
 
 - Move `preprocessFormData` implementation to `@typed-web/form-utils` (re-exported for backwards compatibility)